Przystosowane_do_kompilacji_w_MSVC_Windows11_cd

+++ RTM/src/BASE/wb_randg.c
@@ -102,12 +102,12 @@ float randnorm()
     if  (iset == 0) 
     {
         do {
-            v1=2.0*randg()-1.0;
-            v2=2.0*randg()-1.0;
+            v1=2.0f*randg()-1.0f;
+            v2=2.0f*randg()-1.0f;
             rsq=v1*v1+v2*v2;
         } while (rsq >= 1.0 || rsq == 0.0);
 
-        fac=sqrt(-2.0*log(rsq)/rsq);
+        fac=(float)sqrt(-2.0*log(rsq)/rsq);
         gset=v1*fac;
         iset=1;
         return v2*fac;
